@@ -1,40 +1,48 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 21
-    buildToolsVersion "21.1.1"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "com.rey.material.demo"
-        minSdkVersion 9
-        targetSdkVersion 21
-        versionCode 5
-        versionName "0.0.5"
-    }
-
-    signingConfigs {
-        release {
-            storeFile file(MATERIAL_KEYSTORE_FILE)
-            storePassword MATERIAL_KEYSTORE_PASSWORD
-            keyAlias MATERIAL_KEYSTORE_ALIAS
-            keyPassword MATERIAL_KEYSTORE_PASSWORD
-        }
+        minSdkVersion 14
+        targetSdkVersion 28
+        versionCode 7
+        versionName "0.0.7"
     }
 
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            signingConfig signingConfigs.release
         }
+    }
+
+    if(project.hasProperty('MATERIAL_KEYSTORE_FILE')) {
+        android.signingConfigs {
+            release {
+                storeFile file(MATERIAL_KEYSTORE_FILE)
+                storePassword MATERIAL_KEYSTORE_PASSWORD
+                keyAlias MATERIAL_KEYSTORE_ALIAS
+                keyPassword MATERIAL_KEYSTORE_PASSWORD
+            }
+        }
+
+        android.buildTypes.release.signingConfig signingConfigs.release
     }
 }
 
+ext{
+    androidXVersion = '1.0.0'
+}
+
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:22.2.0'
-    compile 'com.android.support:cardview-v7:22.2.0'
-    compile 'com.squareup.picasso:picasso:2.5.0'
-    compile 'com.squareup.leakcanary:leakcanary-android:1.3.1'
-    compile project(':lib')
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation "androidx.appcompat:appcompat:${androidXVersion}"
+    implementation "androidx.cardview:cardview:${androidXVersion}"
+    implementation "androidx.recyclerview:recyclerview:${androidXVersion}"
+    implementation 'com.squareup.picasso:picasso:2.5.0'
+    implementation 'com.squareup.leakcanary:leakcanary-android:1.3.1'
+    implementation project(':material')
 }
