@@ -16,7 +16,7 @@ import android.os.Build;
 import android.os.Parcel;
 import android.os.Parcelable;
 import android.provider.ContactsContract;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.Editable;
 import android.text.Selection;
 import android.text.SpannableStringBuilder;
@@ -24,7 +24,6 @@ import android.text.Spanned;
 import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
@@ -48,7 +47,6 @@ import com.squareup.picasso.Picasso;
 import com.squareup.picasso.Target;
 
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashMap;
 
 /**
@@ -60,7 +58,7 @@ public class ContactEditText extends EditText{
 
     private HashMap<String, Recipient> mRecipientMap;
 
-    private int mDefaultAvatarId = R.drawable.ic_user;
+    private int mDefaultAvatarId;
     private int mSpanHeight;
     private int mSpanMaxWidth;
     private int mSpanPaddingLeft;
@@ -92,15 +90,12 @@ public class ContactEditText extends EditText{
         super(context, attrs, defStyleAttr);
     }
 
-    public ContactEditText(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
-        super(context, attrs, defStyleAttr);
-    }
-
     @Override
     protected void init(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes){
         mRecipientMap = new HashMap<>();
         mAutoCompleteMode = AUTOCOMPLETE_MODE_MULTI;
 
+        mDefaultAvatarId = R.drawable.ic_user;
         mSpanHeight = ThemeUtil.dpToPx(context, 32);
         mSpanMaxWidth = ThemeUtil.dpToPx(context, 150);
         mSpanPaddingLeft = ThemeUtil.dpToPx(context, 8);
