@@ -16,8 +16,9 @@ import android.graphics.Shader;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.support.annotation.NonNull;
-import android.support.v4.view.GravityCompat;
+import androidx.annotation.NonNull;
+import androidx.core.view.GravityCompat;
+
 import android.text.TextUtils;
 import android.util.AttributeSet;
 import android.util.TypedValue;
@@ -159,7 +160,7 @@ public class ContactView extends FrameLayout implements Target{
     @Override
     public boolean onTouchEvent(@NonNull MotionEvent event) {
         boolean result = super.onTouchEvent(event);
-        return  mRippleManager.onTouchEvent(event) || result;
+        return  mRippleManager.onTouchEvent(this, event) || result;
     }
 
     public void setAvatarBitmap(Bitmap bm){
