@@ -2,16 +2,16 @@ package com.rey.material.demo;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.view.ViewPager;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarActivity;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Gravity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
+import androidx.viewpager.widget.ViewPager;
+
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -29,7 +29,7 @@ import com.rey.material.drawable.ThemeDrawable;
 import com.rey.material.util.ThemeUtil;
 import com.rey.material.util.ViewUtil;
 import com.rey.material.widget.SnackBar;
-import com.rey.material.widget.TabPageIndicator;
+import com.rey.material.widget.TabIndicatorView;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
@@ -40,7 +40,7 @@ public class MainActivity extends AppCompatActivity implements ToolbarManager.On
 	private FrameLayout fl_drawer;
 	private ListView lv_drawer;
 	private CustomViewPager vp;
-	private TabPageIndicator tpi;
+	private TabIndicatorView tiv;
 	
 	private DrawerAdapter mDrawerAdapter;
 	private PagerAdapter mPagerAdapter;
@@ -56,20 +56,20 @@ public class MainActivity extends AppCompatActivity implements ToolbarManager.On
 		super.onCreate(savedInstanceState);
 		
 		setContentView(R.layout.activity_main);
-				
-		dl_navigator = (DrawerLayout)findViewById(R.id.main_dl);
-		fl_drawer = (FrameLayout)findViewById(R.id.main_fl_drawer);
-		lv_drawer = (ListView)findViewById(R.id.main_lv_drawer);
-		mToolbar = (Toolbar)findViewById(R.id.main_toolbar);
-		vp = (CustomViewPager)findViewById(R.id.main_vp);
-		tpi = (TabPageIndicator)findViewById(R.id.main_tpi);
-        mSnackBar = (SnackBar)findViewById(R.id.main_sn);
+
+        lv_drawer = findViewById(R.id.main_lv_drawer);
+		dl_navigator = findViewById(R.id.main_dl);
+		fl_drawer = findViewById(R.id.main_fl_drawer);
+        mToolbar = findViewById(R.id.main_toolbar);
+		vp = findViewById(R.id.main_vp);
+		tiv = findViewById(R.id.main_tiv);
+        mSnackBar = findViewById(R.id.main_sn);
 
         mToolbarManager = new ToolbarManager(getDelegate(), mToolbar, R.id.tb_group_main, R.style.ToolbarRippleStyle, R.anim.abc_fade_in, R.anim.abc_fade_out);
         mToolbarManager.setNavigationManager(new ToolbarManager.ThemableNavigationManager(R.array.navigation_drawer, getSupportFragmentManager(), mToolbar, dl_navigator) {
             @Override
             public void onNavigationClick() {
-                if(mToolbarManager.getCurrentGroup() != R.id.tb_group_main)
+                if (mToolbarManager.getCurrentGroup() != R.id.tb_group_main)
                     mToolbarManager.setCurrentGroup(R.id.tb_group_main);
                 else
                     dl_navigator.openDrawer(GravityCompat.START);
@@ -93,8 +93,8 @@ public class MainActivity extends AppCompatActivity implements ToolbarManager.On
 		
 		mPagerAdapter = new PagerAdapter(getSupportFragmentManager(), mItems);
 		vp.setAdapter(mPagerAdapter);
-		tpi.setViewPager(vp);
-		tpi.setOnPageChangeListener(new ViewPager.OnPageChangeListener() {
+        tiv.setTabIndicatorFactory(new TabIndicatorView.ViewPagerIndicatorFactory(vp));
+		vp.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
 
             @Override
             public void onPageSelected(int position) {
@@ -142,6 +142,11 @@ public class MainActivity extends AppCompatActivity implements ToolbarManager.On
                 mToolbarManager.setCurrentGroup(R.id.tb_group_main);
                 break;
             case R.id.tb_theme:
+                if(mToolbarManager.isNavigationVisisble())
+                    mToolbarManager.setNavigationVisisble(false, true);
+                else
+                    mToolbarManager.setNavigationVisisble(true, true);
+
                 int theme = (ThemeManager.getInstance().getCurrentTheme() + 1) % ThemeManager.getInstance().getThemeCount();
                 ThemeManager.getInstance().setCurrentTheme(theme);
                 Toast.makeText(this, "Current theme: " + theme, Toast.LENGTH_SHORT).show();
@@ -275,7 +280,7 @@ public class MainActivity extends AppCompatActivity implements ToolbarManager.On
 		static {
 			Field f = null;
 			try {
-				Class<?> c = Class.forName("android.support.v4.app.FragmentManagerImpl");
+				Class<?> c = Class.forName("androidx.fragment.app.FragmentManagerImpl");
 				f = c.getDeclaredField("mActive");
 				f.setAccessible(true);   
 			} catch (Exception e) {}
