@@ -1,9 +1,10 @@
 package com.rey.material.demo;
 
 import android.content.Context;
-import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
 import android.view.View;
+
+import androidx.viewpager.widget.ViewPager;
 
 import com.rey.material.widget.Slider;
 import com.rey.material.widget.Switch;
@@ -11,7 +12,7 @@ import com.rey.material.widget.Switch;
 /**
  * Created by Rey on 3/18/2015.
  */
-public class CustomViewPager extends ViewPager{
+public class CustomViewPager extends ViewPager {
 
     public CustomViewPager(Context context, AttributeSet attrs) {
         super(context, attrs);
