@@ -1,14 +1,14 @@
 package com.rey.material.demo;
 
 import android.annotation.TargetApi;
-import android.content.Context;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
+
+import androidx.fragment.app.Fragment;
 
 import com.rey.material.app.BottomSheetDialog;
 import com.rey.material.app.DatePickerDialog;
